+++ src/connected/AppPane.tsx
@@ -39,7 +39,7 @@ const AppPane: React.FC<AppPaneProps> = ({classes}) => {
     // <Toolbar/>: Empty toolbar is a spacer, see docs https://material-ui.com/components/app-bar/
     return (
         <main className={classes.main}>
-            <Toolbar/>
+            <Toolbar variant="dense"/>
             <ControlBar/>
             <Workspace/>
         </main>